requirements.txt
pandas 
numpy
sentence-transformers
faiss-cpu
streamlit==1.37.0